LPS-160259 Also move tests to corresponding module
